+++ huey/__init__.py
@@ -17,6 +17,7 @@ from huey.api import chord
 from huey.api import crontab
 from huey.api import group
 from huey.exceptions import CancelExecution
+from huey.exceptions import ResultMissing
 from huey.exceptions import RetryTask
 from huey.utils import Error
 from huey.utils import SKIPPED